moved function to find last valid value in array to abiotic tools
