+++ tests/models/abiotic/test_wind.py
@@ -1,7 +1,6 @@
 """Test module for abiotic.wind.py."""
 
 import numpy as np
-import pytest
 
 from virtual_ecosystem.core.constants import CoreConsts
 from virtual_ecosystem.models.abiotic.constants import AbioticConsts
@@ -181,34 +180,6 @@ def test_generate_relative_turbulence_intensity():
     )
     np.testing.assert_allclose(result_t, exp_result_t, rtol=1e-3, atol=1e-3)
     np.testing.assert_allclose(result_f, exp_result_f, rtol=1e-3, atol=1e-3)
-
-
-@pytest.mark.parametrize(
-    "input_array, expected",
-    [
-        (np.array([[1.0, 2.0, 3.0], [4.0, 5.0, 6.0]]), np.array([4.0, 5.0, 6.0])),
-        (
-            np.array([[1.0, np.nan, 3.0], [4.0, 5.0, np.nan], [np.nan, 8.0, 9.0]]),
-            np.array([4.0, 8.0, 9.0]),
-        ),
-        (
-            np.array([[np.nan, 2.0, np.nan], [np.nan, 5.0, np.nan]]),
-            np.array([np.nan, 5.0, np.nan]),
-        ),
-        (np.array([[np.nan, 2.0, 3.0]]), np.array([np.nan, 2.0, 3.0])),
-        (
-            np.array([[np.nan, np.nan, np.nan], [np.nan, np.nan, np.nan]]),
-            np.array([np.nan, np.nan, np.nan]),
-        ),
-    ],
-)
-def test_find_last_valid_row(input_array, expected):
-    """Test that last true value is selected for each column."""
-
-    from virtual_ecosystem.models.abiotic.wind import find_last_valid_row
-
-    result = find_last_valid_row(input_array)
-    np.testing.assert_allclose(result, expected)
 
 
 def test_calculate_wind_attenuation_coefficient(dummy_climate_data):
